# packages/api/src/openapi/swagger.yaml
openapi: 3.0.0
info:
  title: Royal Burger API
  description: API para o sistema de delivery da hamburgueria Royal Burger.
  version: "1.0.0"
servers:
  - url: http://127.0.0.1:5000/api
    description: Servidor de Desenvolvimento Local

# ===============================================================
# Componentes Reutilizáveis
# ===============================================================
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponse:
      type: object
      properties:
        msg:
          type: string
          example: "Operação realizada com sucesso."

    # --- Schemas para Chat ---
...
    Customer:
      type: object
      properties:
        id: {type: integer, example: 3}
        full_name: {type: string, example: "João da Silva"}
        email: {type: string, format: email, example: "joao.silva@email.com"}
        phone_number: {type: string, example: "14999998888"}
        role: {type: string, example: "customer"}
        is_active: {type: boolean, example: true}
    NewCustomer:
      type: object
      required:
        - full_name
        - email
        - password
      properties:
        full_name: {type: string, example: "Maria Oliveira"}
        email: {type: string, format: email, example: "maria.oliveira@email.com"}
        password: {type: string, example: "SenhaForte123"}
        phone_number: {type: string, example: "14988887777"}

    # --- Schemas para Endereços (Address) ---
    Address:
      type: object
      properties:
        id: {type: integer, example: 1}
        street: {type: string, example: "Rua das Flores"}
        number: {type: string, example: "123"}
        complement: {type: string, example: "Apto 4B"}
        neighborhood: {type: string, example: "Centro"}
        city: {type: string, example: "Lins"}
        state: {type: string, example: "SP"}
        zip_code: {type: string, example: "16400-000"}
    NewAddress:
      type: object
      required:
        - street
        - number
        - neighborhood
        - city
      properties:
        street: {type: string, example: "Avenida Brasil"}
        number: {type: string, example: "1000"}
        complement: {type: string}
        neighborhood: {type: string, example: "Jardim Americano"}
        city: {type: string, example: "Lins"}
        state: {type: string, example: "SP"}
        zip_code: {type: string, example: "16400-123"}

    # --- Schemas para Fidelidade (Loyalty) ---
    LoyaltyBalance:
      type: object
      properties:
        accumulated_points: {type: integer, example: 500}
        spent_points: {type: integer, example: 100}
        current_balance: {type: integer, example: 400}
    LoyaltyHistoryEntry:
      type: object
      properties:
        points: {type: integer, example: 50}
        reason: {type: string, example: "Pedido ID: 123"}
        date: {type: string, format: date-time, example: "2025-09-15T20:30:00"}

    # --- Schemas para Ingredientes ---
...
security:
  - bearerAuth: []

# ===============================================================
# Paths (Endpoints da API)
# ===============================================================
paths:
...
  /customers:
    post:
      summary: Cria um novo cliente (cadastro).
      security: []
      tags: [Clientes]
      requestBody:
        required: true
        content:
          application/json: {schema: {$ref: '#/components/schemas/NewCustomer'}}
      responses:
        '201':
          description: Cliente criado com sucesso.
          content:
            application/json: {schema: {$ref: '#/components/schemas/Customer'}}
    get:
      summary: (Admin) Lista todos os clientes.
      tags: [Clientes]
      responses:
        '200':
          description: Lista de clientes.
          content:
            application/json: {schema: {type: array, items: {$ref: '#/components/schemas/Customer'}}}
  /customers/{user_id}:
    get:
      summary: Busca dados de um cliente específico.
      tags: [Clientes]
      parameters:
        - {name: user_id, in: path, required: true, schema: {type: integer}}
      responses:
        '200':
          description: Dados do cliente.
          content:
            application/json: {schema: {$ref: '#/components/schemas/Customer'}}
    put:
      summary: Atualiza os dados de um cliente.
      tags: [Clientes]
      parameters:
        - {name: user_id, in: path, required: true, schema: {type: integer}}
      requestBody:
        content:
          application/json: {schema: {$ref: '#/components/schemas/NewCustomer'}}
      responses:
        '200':
          description: Dados atualizados com sucesso.
          content:
            application/json: {schema: {$ref: '#/components/schemas/MessageResponse'}}
    delete:
      summary: Inativa a conta de um cliente.
      tags: [Clientes]
      parameters:
        - {name: user_id, in: path, required: true, schema: {type: integer}}
      responses:
        '200':
          description: Conta inativada com sucesso.
          content:
            application/json: {schema: {$ref: '#/components/schemas/MessageResponse'}}

  # --- Rotas de Endereços ---
  /customers/{user_id}/addresses:
    get:
      summary: Lista os endereços de um cliente.
      tags: [Endereços]
      parameters:
        - {name: user_id, in: path, required: true, schema: {type: integer}}
      responses:
        '200':
          description: Lista de endereços.
          content:
            application/json: {schema: {type: array, items: {$ref: '#/components/schemas/Address'}}}
    post:
      summary: Adiciona um novo endereço para um cliente.
      tags: [Endereços]
      parameters:
        - {name: user_id, in: path, required: true, schema: {type: integer}}
      requestBody:
        required: true
        content:
          application/json: {schema: {$ref: '#/components/schemas/NewAddress'}}
      responses:
        '201':
          description: Endereço criado com sucesso.
          content:
            application/json: {schema: {$ref: '#/components/schemas/Address'}}
  /customers/addresses/{address_id}:
    put:
      summary: Atualiza um endereço específico.
      tags: [Endereços]
      parameters:
        - {name: address_id, in: path, required: true, schema: {type: integer}}
      requestBody:
        required: true
        content:
          application/json: {schema: {$ref: '#/components/schemas/NewAddress'}}
      responses:
        '200':
          description: Endereço atualizado.
          content:
            application/json: {schema: {$ref: '#/components/schemas/MessageResponse'}}
    delete:
      summary: Deleta um endereço específico.
      tags: [Endereços]
      parameters:
        - {name: address_id, in: path, required: true, schema: {type: integer}}
      responses:
        '200':
          description: Endereço deletado.
          content:
            application/json: {schema: {$ref: '#/components/schemas/MessageResponse'}}
            
  # --- Rotas de Fidelidade ---
  /customers/{user_id}/loyalty/balance:
    get:
      summary: Consulta o saldo de pontos de fidelidade.
      tags: [Fidelidade]
      parameters:
        - {name: user_id, in: path, required: true, schema: {type: integer}}
      responses:
        '200':
          description: Saldo de pontos.
          content:
            application/json: {schema: {$ref: '#/components/schemas/LoyaltyBalance'}}
  /customers/{user_id}/loyalty/history:
    get:
      summary: Consulta o histórico de pontos de fidelidade.
      tags: [Fidelidade]
      parameters:
        - {name: user_id, in: path, required: true, schema: {type: integer}}
      responses:
        '200':
          description: Histórico de transações de pontos.
          content:
            application/json: {schema: {type: array, items: {$ref: '#/components/schemas/LoyaltyHistoryEntry'}}}

  # --- Rotas de Ingredientes (Gerenciamento) ---
...